+++ src/app/controllers/ModelosMensagensController.js
@@ -21,7 +21,7 @@ class ModelosMensagensController {
   }
 
   async store(req, res) {
-    await validateSchema(req.body, MODELO_SCHEMA, res);
+    if (!(await validateSchema(req.body, MODELO_SCHEMA, res))) return;
 
     const { title, greeting, content } = req.body;
 
@@ -36,8 +36,8 @@ class ModelosMensagensController {
   }
 
   async update(req, res) {
-    await validateId(req.params.id, res);
-    await validateSchema(req.body, MODELO_SCHEMA, res);
+    if (!(await validateId(req.params.id, res))) return;
+    if (!(await validateSchema(req.body, MODELO_SCHEMA, res))) return;
 
     const { title, greeting, content } = req.body;
 
@@ -52,7 +52,7 @@ class ModelosMensagensController {
   }
 
   async delete(req, res) {
-    await validateId(req.params.id, res);
+    if (!(await validateId(req.params.id, res))) return;
 
     const nDeleted = await ModeloMensagem.destroy({
       where: { id: req.params.id, id_escola: req.userId },
